🚚rename: RoundResult 디렉토리 변경(domain → dto)
