+++ src/main/java/racingcar/dto/RoundResult.java
@@ -1,6 +1,8 @@
-package racingcar.domain;
+package racingcar.dto;
 
 import java.util.List;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 public record RoundResult(CarsDto carsDto) {
