+++ src/test/java/racingcar/domain/GameTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.dto.RoundResult;
 import racingcar.util.NumberGenerator;
 
 @DisplayName("Game 단위 테스")
